# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API circleoflinks",
        "description": "API de red social de profecionales ",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/circleoflinks-app/public"
        }
    ],
    "paths": {
...
        "/api/v1/statusofertalaboral": {
            "get": {
                "tags": [
                    "Status Oferta laboral"
                ],
                "summary": "Obtener todos los registros de una oferta laboral",
                "description": "Devuelve información detallada sobre una status.",
                "operationId": "0325ff1e06e02b7734be1767db74d9c2",
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "closed"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "status Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/statusofertalaboral/{id}": {
            "get": {
                "tags": [
                    "Status Oferta laboral"
                ],
                "summary": "Obtener registro especifico de un status oferta laboral",
                "description": "Devuelve información detallada sobre un status.",
                "operationId": "a0df800e119e79b2c0dfb8029ac995f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Oferta laboral encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "closed"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": " oferta laboral cerrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Oferta laboral no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "status Oferta laboral no encontrada con ID: {id}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}